Fixed a warning (int and unsigned int comparison) in Camera.cpp Updated example to switch between LSH and KDTree depending on descriptors type detected. Also use SIFT as it gives better results than SURF.
git-svn-id: http://find-object.googlecode.com/svn/trunk/find_object@163 620bd6b2-0a58-f614-fd9a-1bd335dccda9

+++ src/Camera.cpp
@@ -83,7 +83,7 @@ void Camera::takeImage()
 	}
 	else if(!images_.empty())
 	{
-		if(currentImageIndex_ < images_.size())
+		if(currentImageIndex_ < (unsigned int)images_.size())
 		{
 			img = cv::imread(images_[currentImageIndex_++]);
 		}
